Move getSectionMetadata
