+++ libs/utils/utils.js
@@ -568,15 +568,3 @@ export function createIntersectionObserver({ el, callback, once = true, options 
   io.observe(el);
   return io;
 }
-
-export const getSectionMetadata = (el) => {
-  if (!el) return {};
-  const metadata = {};
-  el.childNodes.forEach((node) => {
-    const key = node.children?.[0]?.textContent?.toLowerCase();
-    if (!key) return;
-    const val = node.children?.[1]?.textContent?.toLowerCase();
-    metadata[key] = val;
-  });
-  return metadata;
-};
