+++ test/blocks/section-metadata/section-meta.test.js
@@ -5,7 +5,7 @@ import { readFile } from '@web/test-runner-commands';
 import { expect } from '@esm-bundle/chai';
 
 document.body.innerHTML = await readFile({ path: './mocks/body.html' });
-const { default: init } = await import('../../../libs/blocks/section-metadata/section-metadata.js');
+const { default: init, getSectionMetadata } = await import('../../../libs/blocks/section-metadata/section-metadata.js');
 
 describe('Section Metdata', () => {
   it('Gracefully dies', () => {
@@ -27,5 +27,10 @@ describe('Section Metdata', () => {
     const sm = sec.querySelector('.section-metadata');
     init(sm);
     expect(sec.style.backgroundColor).to.equal('rgb(239, 239, 239)');
+  });
+
+  it('gets section metadata', () => {
+    const expected = { style: 'darkest, xxl spacing', background: 'rgb(239, 239, 239)' };
+    expect(getSectionMetadata(document.querySelector('.section.color .section-metadata'))).to.eql(expected);
   });
 });
